requirements.txt
# Python dependency pins for the K* verification suite.
# Upper bounds are important: they prevent a future breaking change
# in a transitive dep from silently invalidating the verification.
# See .zenodo.json for the Lean-side pin (lean4/lean-toolchain).

numpy>=1.24,<3
sympy>=1.12,<2
scipy>=1.10,<2
cvxpy>=1.4,<2
pyyaml>=6.0,<7
highspy>=1.7,<2         # tier3-numpy basis-cover ILP certification
pulp>=2.7,<3            # tier3-numpy CBC cross-solver round-trip

# Optional: notebook deps (use notebook/requirements.txt instead
# if you only need the Jupyter demo, not the full verification suite).
# pandas>=2.0,<4
# jupyter>=1.0,<2
